Refactor and silence warnings
